proyecto/app/requirements.txt
Flask==1.1.1
cassandra-driver==3.23.0
